Fix CI: Windows %zu format + mobile/web test bundling
- image.c: cast size_t to unsigned long long for portable format (MinGW's
  -Werror rejects %zu).
- tests: gate filesystem-fixture tests (test_image, test_obj_loader) to
  desktop-only builds. Android APK, iOS .app, and Emscripten virtual FS
  don't expose the host source tree, so SDL3D_TEST_ASSETS_DIR is
  meaningless there.

+++ src/image.c
@@ -70,7 +70,7 @@ bool sdl3d_load_image_from_memory(const void *data, size_t size, sdl3d_image *ou
     }
     if (size == 0 || size > (size_t)INT_MAX)
     {
-        return SDL_SetError("Image buffer size %zu is out of range for stbi.", size);
+        return SDL_SetError("Image buffer size %llu is out of range for stbi.", (unsigned long long)size);
     }
 
     int w = 0;
